+++ src/app/providers/detectorProvider.tsx
@@ -1,8 +1,8 @@
 import createActivityDetector from 'activity-detector';
 import { ReactNode, useEffect } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { useAuthStore } from '@/shared/model/authStore';
-import { useDetectorStore } from '@/shared/model/detectorStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
+import { useDetectorStore } from '@/shared/model/store/detectorStore';
 import { userQueries } from '@/features/user/api/queries';
 import { LOGIN_STATUS } from '@/entities/user/model/constants';
