+++ src/features/auth/ui/SignOutButton.tsx
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { useAuthStore } from '@/shared/model/authStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
 import { ROUTES } from '@/shared/constants/routes';
 import { authQueries } from '../api/queries';
