+++ src/features/auth/ui/WithdrawModal.tsx
@@ -2,9 +2,9 @@ import { useForm } from 'react-hook-form';
 import { ErrorMessage } from '@hookform/error-message';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
-import { useAuthStore } from '@/shared/model/authStore';
-import { useModalStore } from '@/shared/model/modalStore';
-import { useSocketStore } from '@/shared/model/socketStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
+import { useModalStore } from '@/shared/model/store/modalStore';
+import { useSocketStore } from '@/shared/model/store/socketStore';
 import { MODAL } from '@/shared/constants/modal';
 import { ROUTES } from '@/shared/constants/routes';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
