+++ src/features/directMessage/ui/DMView.tsx
@@ -2,10 +2,10 @@ import { invariant } from 'es-toolkit/compat';
 import { useParams } from 'react-router-dom';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { useChatStore } from '@/shared/model/chatStore';
-import { useDetectorStore } from '@/shared/model/detectorStore';
-import { useSocketStore } from '@/shared/model/socketStore';
-import { useUnreadMessagesStore } from '@/shared/model/unreadMessagesStore';
+import { useChatStore } from '@/shared/model/store/chatStore';
+import { useDetectorStore } from '@/shared/model/store/detectorStore';
+import { useSocketStore } from '@/shared/model/store/socketStore';
+import { useUnreadMessagesStore } from '@/shared/model/store/unreadMessagesStore';
 import { useInfiniteScroll } from '@/shared/lib/hooks/useInfiniteScroll';
 import { DMQueries } from '../api/queries';
 import { groupMessages } from '../lib/utils';
