+++ src/features/directMessage/ui/MessageGroup.tsx
@@ -1,5 +1,5 @@
 import type { MessageGroups } from '../model/types';
-import { useAuthStore } from '@/shared/model/authStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
 import { DiscordIcon } from '@/shared/icons/DiscordIcon';
 import { formatMessageDate } from '@/shared/lib/utils/formatMessageDate';
 import { cn } from '@/shared/lib/utils/utils';
