+++ src/features/server/ui/ChannelCategories.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { useModalStore } from '@/shared/model/modalStore';
+import { useModalStore } from '@/shared/model/store/modalStore';
 import { MODAL } from '@/shared/constants/modal';
 import { PlusIcon } from '@/shared/icons/PlusIcon';
 import { EmptyList } from '@/shared/ui/EmptyList';
