+++ src/features/server/ui/InvitedMemberModal.tsx
@@ -3,7 +3,7 @@ import { useLocation, useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { useState } from 'react';
 import { useInfiniteQuery, useMutation, useQuery } from '@tanstack/react-query';
-import { useModalStore } from '@/shared/model/modalStore';
+import { useModalStore } from '@/shared/model/store/modalStore';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
 import { MODAL } from '@/shared/constants/modal';
