+++ src/features/user/ui/UserProfileModal.tsx
@@ -2,8 +2,8 @@ import { toast } from 'react-toastify';
 import { useEffect, useState } from 'react';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import type { UserProfile } from '@/entities/user/model/types';
-import { useAuthStore } from '@/shared/model/authStore';
-import { useModalStore } from '@/shared/model/modalStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
+import { useModalStore } from '@/shared/model/store/modalStore';
 import { userQueries } from '@/entities/user/api/queries';
 import { MODAL } from '@/shared/constants/modal';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
