+++ src/shared/api/interceptors.ts
@@ -1,6 +1,6 @@
 import type { AxiosError, AxiosResponse, InternalAxiosRequestConfig } from 'axios';
 import type { ApiErrorResponse } from '../types/apiResponse';
-import { useAuthStore } from '@/shared/model/authStore';
+import { useAuthStore } from '@/shared/model/store/authStore';
 import { NetworkOfflineError, TokenExpiredHandler } from './errorHandler';
 
 export const requestInterceptor = (config: InternalAxiosRequestConfig) => {
