+++ src/shared/model/store/detectorStore.ts
