+++ src/shared/model/store/serverStore.ts
