+++ src/shared/ui/ChannelMessageDefault.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
-import { useModalStore } from '../model/modalStore';
+import { useModalStore } from '../model/store/modalStore';
 import { MODAL } from '../constants/modal';
 import { ArrowRight } from '../icons/ArrowRight';
 import { NewMemberIcon } from '../icons/NewMemberIcon';
